Routes, displaying Products

+++ database/migrations/2021_11_09_083132_create_menus_table.php
@@ -18,7 +18,7 @@ class CreateMenusTable extends Migration
             $table->string('sq_menu_title', 255)->index();
             $table->string('en_menu_title', 255)->index();
             $table->string('route_name', 255);
-            $table->integer('parent_id')->default(0);
+            $table->bigInteger('parent_id')->default(0);
             $table->string('sort_order')->default(0);
             $table->string('slug', 255);
             $table->timestamps();
